Rename folder from NoteList to Notes.
