+++ src/components/Notes/style.css
