+++ src/index.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import Routes from './routes';
-
 import './index.css';
 
 ReactDOM.render(
